restructure to remove intermingling of function folders and non function folders
