+++ app/__init__.py
@@ -1,0 +1,4 @@
+'''
+Init the app module
+This module is intended for code relation to app behavior that will not be an Azure Function
+'''
