+++ app/decorators/__init__.py
@@ -1,0 +1,2 @@
+'''Init the decorators module'''
+from app.decorators.error_aggregation import handle_error as catcher
