+++ app/services/__init__.py
@@ -1,0 +1,6 @@
+'''Init the service module'''
+
+from app.services import card as CardService
+from app.services import event as EventService
+from app.services import game as GameService
+from app.services import user as UserService
