--- decorators/__init__.py
@@ -1,2 +1,0 @@
-'''Init the decorators module'''
-from decorators.exception_aggregation import handle_exception as catcher
