--- services/__init__.py
@@ -1,6 +1,0 @@
-'''Init the service module'''
-
-from services import card as CardService
-from services import event as EventService
-from services import game as GameService
-from services import user as UserService
